additional fixes by cursor
