+++ src/contentScripts/views/DrupalCoreVersions.ts
